Add real time fundraiser updates
* The app now increments the amount raised in real time using a
combination of a scheduled job (per app installation) and a real time
message channel. The scheduled job checks for posts that need updates
(haven't reached their fundraiser end date) and then fetches new raised
details for them and sends them via the real time channel to the
subscribed custom post. The custom post uses useState for all of the UGC
and fundraiser values, and it uses the state setters along with the
updated data via its channel subscription to set updated values for the
elements of the returned jsx, namely the amount raised. It also fills
the progress bar.
* There are a bunch of logging messages I need to remove that I put
there while getting things to work
* Broke out a bunch of functionality into separate files and even
classes. There's more to be done on that front, though.